+++ app/controllers/public/comics_controller.rb
@@ -8,7 +8,7 @@ class Public::ComicsController < ApplicationController
     @comic = Comic.find(params[:id])
     @review = Review.find(params[:id])
     @reviews = @comic.reviews
-    @tag_list = @review.tags
+    @tag = Tag.new
   end
 
 end
