--- app/controllers/public/tag_maps_controller.rb
@@ -1,9 +1,0 @@
-class Public::TagMapsController < ApplicationController
-
-  def create
-  end
-
-  def destroy
-  end
-
-end
