+++ app/models/comic.rb
@@ -2,6 +2,7 @@ class Comic < ApplicationRecord
 
   belongs_to :category
   has_many :reviews, dependent: :destroy
+  has_many :tags, dependent: :destroy
 
   attachment :comic_image
