--- app/models/tag_map.rb
@@ -1,9 +1,0 @@
-class TagMap < ApplicationRecord
-
-  belongs_to :tag
-  belongs_to :review
-
-  validates :review_id, presence: true
-  validates :tag_id, presence: true
-
-end
