+++ app/views/public/tags/search.html.erb
