--- db/migrate/20210108043532_create_tag_maps.rb
@@ -1,10 +1,0 @@
-class CreateTagMaps < ActiveRecord::Migration[5.2]
-  def change
-    create_table :tag_maps do |t|
-      t.references :review, foreign_key: true
-      t.references :tag, foreign_key: true
-
-      t.timestamps
-    end
-  end
-end
